% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{load_input_data}
\alias{load_input_data}
\title{Load data}
\usage{
load_input_data(
  filename,
  schema = c("Zeit", "Jahr", "Monat", "Tag", "HR", "K", "A1", "A2", "A3", "A4", "A5",
    "A6", "A7", "MW_J")
)
}
\arguments{
\item{filename}{the filename on disk containing the data}

\item{schema}{- an optional parameters holding the data schema. 
The current default value corresponds to the actual processing code}
}
\value{
data - the in-memory object containg the data
}
\description{
loading data from a csv file
}
\examples{
data  <- load_input_data(filename = file.choose)
data <- load_input_data(filename = file.path("tests/input/sample_complete_10y.csv"))

}
