% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{parameters}
\alias{parameters}
\title{Set analysis parameters}
\usage{
parameters(
  filename = "./input/params.csv",
  schema = c("d", "e", "FL_d", "FX_h", "AlphaParameter", "Anzahl.unmoeglich.max",
    "Anzahl.leer.wdh.max", "Anzahl.leer.gesamt.max", "Anzahl.Wertwdh.Null.max",
    "Anzahl.Wertwdh.Wert.max", "Anzahl.unmoeglich.hist.max", "Anzahl.leer.wdh.hist.max",
    "Anzahl.leer.gesamt.hist.max", "Anzahl.Wertwdh.Null.hist.max",
    "Anzahl.Wertwdh.Wert.hist.max", "Anzahl.Wertwdh.Wert.tol.hist.max")
)
}
\arguments{
\item{filename}{an optional paramneter, the filename on disk containing the data. If missing, the default values are loaded}

\item{schema}{- an optional parameters holding the data schema. The current default value corresponds to the actual processing code}
}
\value{
the in-memory object containg the data
}
\description{
loads analysis parameters from a data file or return default values
}
\examples{
parameters  <- parameters()
parameters <- parameters(filename = file.path("tests/input/params.csv"))

}
