% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_charts}
\alias{plot_charts}
\title{Draw specified charts}
\usage{
plot_charts(names = names, datapath = datapath, i18n = i18n)
}
\arguments{
\item{names}{the names of the generated data resources, to be used for plotting
For example, if user language is `"en"` then the names look like `"Chart1"`, `"Chart2"`, ... 
corresponding to existent data files `"Chart1.csv"`, `"Chart2.csv"`, ...
If the user language is `"de"` then names are like `"Grafik1"`, ... corresponding to `"Grafik1.csv"`, ... data}

\item{datapath}{- optional, the path on disk where all generated data will be saved}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a list of plotly objects
}
\description{
Draw specified charts
}
\examples{
plots <- plot_charts(names = chartnames, datapath = file.path("tests/output"), i18n = i18n)

}
