% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Descriptive_Statistics}
\alias{plot_Descriptive_Statistics}
\title{Descriptive_Statistics}
\usage{
plot_Descriptive_Statistics(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Descriptive_Statistics 
Requires the following data columns:
"Jahr",
"me_HR_J", "me_K_J", "me_A1_J", "me_A2_J", "me_A3_J", "me_A4_J", "me_A5_J", "me_A6_J", "me_A7_J",
"sd_HR_J", "sd_K_J", "sd_A1_J", "sd_A2_J", "sd_A3_J", "sd_A4_J", "sd_A5_J", "sd_A6_J", "sd_A7_J"
}
\examples{
p1 <- plot_Descriptive_Statistics(name = i18n$t("Descriptive_Statistics"), datapath= file.path("tests/output"), i18n = i18n)

}
