% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Fehlermasse}
\alias{plot_Fehlermasse}
\title{Fehlermasse}
\usage{
plot_Fehlermasse(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Fehlermasse 
Requires the following data columns:
"AM", "K", "BTU", "A1", "A2", "A3", "A4", "A5", "A6", "A7"
Row names are 
"mse", "mae", "me"
}
\examples{
p1 <- plot_Fehlermasse(name = i18n$t("Error_Mass"), datapath = file.path("tests/output"), i18n = i18n)

}
