% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Fehlermasse_J}
\alias{plot_Fehlermasse_J}
\title{Fehlermasse_J - UNDE CONSTRUCTION}
\usage{
plot_Fehlermasse_J(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Fehlermasse_J 
Requires the following data columns:
"Jahr",
"installierte Leistung [MW]", "mittlere Einspeisung [MW]",
"Anbieteranzahl",
"mse_AM_J", "mse_K_J", "mse_BTU_J", "mse_A1_J", "mse_A2_J", "mse_A3_J", "mse_A4_J", "mse_A5_J", "mse_A6_J", "mse_A7_J"
"mae_AM_J", "mae_K_J", "mae_BTU_J", "mae_A1_J", "mae_A2_J", "mae_A3_J", "mae_A4_J", "mae_A5_J", "mae_A6_J", "mae_A7_J",
"me_AM_J", "me_K_J", "me_BTU_J", "me_A1_J", "me_A2_J", "me_A3_J", "me_A4_J", "me_A5_J", "me_A6_J", "me_A7_J",
}
\examples{
p1 <- plot_Fehlermasse_J(name = i18n$t("Error_Measures_Yearly"), datapath = file.path("tests/output"), i18n = i18n)

}
