% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Fehlermasse_M}
\alias{plot_Fehlermasse_M}
\title{Fehlermasse_M - UNDE CONSTRUCTION}
\usage{
plot_Fehlermasse_M(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Fehlermasse_M 
Requires the following data columns:
"Jahr", "Monat",
"Anbieteranzahl_M",
"mse_AM_M", "mse_K_M", "mse_BTU_M", "mse_A1_M", "mse_A2_M", "mse_A3_M", "mse_A4_M", "mse_A5_M", "mse_A6_M", "mse_A7_M"
"mae_AM_M", "mae_K_M", "mae_BTU_M", "mae_A1_M", "mae_A2_M", "mae_A3_M", "mae_A4_M", "mae_A5_M", "mae_A6_M", "mae_A7_M",
"me_AM_M", "me_K_M", "me_BTU_M", "me_A1_M", "me_A2_M", "me_A3_M", "me_A4_M", "me_A5_M", "me_A6_M", "me_A7_M",
}
\examples{
p1 <- plot_Fehlermasse_M(name = i18n$t("Error_Measures_Monthly"), datapath = file.path("tests/output"), i18n = i18n)

}
