% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Grafik1}
\alias{plot_Grafik1}
\title{Draw Grafik1}
\usage{
plot_Grafik1(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Grafik1 in the initial code  
Requires the following data columns:
"Jahr",
"mse_A1_J","mse_A2_J","mse_A3_J","mse_A4_J","mse_A5_J","mse_A6_J","mse_A7_J",
"MW_J","JMW"
}
\examples{
p1 <- plot_Grafik1(name = i18n$t("Chart1"), datapath = file.path("tests/output"), i18n = i18n)

}
