% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Grafik2}
\alias{plot_Grafik2}
\title{Draw Grafik2}
\usage{
plot_Grafik2(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Grafik2 described in the initial code  
Requires the following data columns:
"Jahr",
"mse_A1_J/JMW","mse_A2_J/JMW","mse_A3_J/JMW","mse_A4_J/JMW","mse_A5_J/JMW","mse_A6_J/JMW","mse_A7_J/JMW"
}
\examples{
p1 <- plot_Grafik2(name = i18n$t("Chart2"), datapath = file.path("tests/output"), i18n = i18n)

}
