% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Grafik3}
\alias{plot_Grafik3}
\title{Draw Grafik3}
\usage{
plot_Grafik3(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Grafik3 described in the initial code  
Requires the following data columns:
"Jahr", "Monat",
"mse_AM","mse_K", "mse_BTU", "Anbieteranzahl"
}
\examples{
p1 <- plot_Grafik3(name = i18n$t("Chart3"), datapath = file.path("tests/output"), i18n = i18n)

}
