% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Grafik4}
\alias{plot_Grafik4}
\title{Draw Grafik4}
\usage{
plot_Grafik4(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Grafik4 described in the initial code  
Requires the following data columns:
"Monat",
"Obere", "Untere", "me_BTU_M", "Anbieteranzahl"
}
\examples{
p1 <- plot_Grafik4(name = i18n$t("Chart4"), datapath = file.path("tests/output"), i18n = i18n)

}
