% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Grafik5}
\alias{plot_Grafik5}
\title{Draw Grafik5 - UNDER CONSTRUCTION}
\usage{
plot_Grafik5(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Grafik5 described in the initial code  
Requires the following data columns:
"Jahr", "Monat",
"Vergleich", "Durchschnitt", "Verbesserung", "Anzahl_Verb", "Verschlechterung", "Anzahl_Vers"
}
\examples{
p1 <- plot_Grafik5(name = i18n$t("Chart5"), datapath = file.path("tests/output"), i18n = i18n)

}
