% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Grafik6}
\alias{plot_Grafik6}
\title{Draw Grafik6 - UNDER CONSTRUCTION}
\usage{
plot_Grafik6(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Grafik6 
Requires the following data columns:
"Jahr", 
"Vergleich", "Durchschnitt", "Verbesserung", "Anzahl", "Verschlechterung", "Anzahl"
}
\examples{
p1 <- plot_Grafik6(name = i18n$t("Chart6"), datapath = file.path("tests/output"), i18n = i18n)

}
