% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Prognosefehler_Gesamt}
\alias{plot_Prognosefehler_Gesamt}
\title{Prognosefehler_Gesamt}
\usage{
plot_Prognosefehler_Gesamt(name, datapath = datapath, i18n = i18n)
}
\arguments{
\item{name}{the name of the file on disk to be used as data source}

\item{datapath}{- optional, the path on disk where data is stored}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\value{
a plotly object
}
\description{
Draw Prognosefehler_Gesamt 
Requires the following data columns:
"Z", 
"Anbieteranzahl",
"AM", "K", "BTU", "PF_AM", "PF_K", "PF_BTU"
}
\examples{
p1 <- plot_Prognosefehler_Gesamt(name = i18n$t("Forecast_Error_Total"), datapath = file.path("tests/output"), i18n = i18n)

}
