% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_forecast.R
\name{run_forecast_BTU}
\alias{run_forecast_BTU}
\title{Run BTU forecast analysis}
\usage{
run_forecast_BTU(
  EingabeZR,
  parameters,
  datapath = file.path("output"),
  i18n = i18n
)
}
\arguments{
\item{EingabeZR}{the filename on disk containing the data}

\item{parameters}{- an optional parameters holding the data schema. The current default value corresponds to the actual processing code}

\item{datapath}{- the path on disk where all generated data will be saved}

\item{i18n}{- the translator object used to personalize output according with user set language}
}
\value{
list(
  "BTU" = BTU,
  "HRN" = HRN,
 "duration" = duration
)
or NA in case of error
}
\description{
Run BTU forecast analysis 
__when use default parameters the script needs data for at least 1 year (35040 15 Minute frames)__
}
\examples{
results <- run_forecast_BTU(raw_data, parameters)
data <- load_input_data(filename = "./tests/input/sample_complete_10y.csv")

}
