% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{save_plots}
\alias{save_plots}
\title{Save  multiple plots into many formats and many resolutions images}
\usage{
save_plots(plotdata = plotdata, formats = formats, resolutions = resolutions)
}
\arguments{
\item{plotdata, }{a list of plotly objects}

\item{formats}{an array of formats e.g. `formats = c("png", "pdf")`}

\item{resolutions}{an array of resolution strings e.g. ` resolutions = c("1920x1440","1920x1200", "1920x1080")`}

\item{i18n}{- required, the translator object used to personalize output according with user set language}
}
\description{
save plots using many formats and many resolutions
supported formats: "png", "jpeg", "pdf", " eps"
supported resolutions:
4:3 aspect ratio resolutions: 1024x768, 1280x960, 1400x1050, 1440x1080 , 1600x1200, 1856x1392, 1920x1440, and 2048x1536.
16:10 aspect ratio resolutions: 1280x800, 1440x900, 1680x1050, 1920x1200 and 2560x1600.
16:9 aspect ratio resolutions: 1024x576, 1152x648, 1280x720, 1366x768, 1600x900, 1920x1080, 2560x1440 and 3840x2160.
}
\examples{
save_plots(plotdata = plots, formats = c("png", "pdf"),  resolutions = c("1920x1440","1920x1200", "1920x1080"))

}
